+++ src/main.c
@@ -9,6 +9,7 @@
 #include "dynawave.h"
 #include "random.h"
 #include "music.h"
+#include "undo.h"
 
 #define DEFAULT_OFFSET 4
 #define ANIM_TIME 8
@@ -21,8 +22,6 @@
 #define PLAYER_GOAL_START 24
 
 #define FIELD_SIZE 256
-#define UNDO_BUFFER_SIZE 256
-#define MAX_UNDO_MOVES (UNDO_BUFFER_SIZE * 2)
 
 char tick = 0;
 
@@ -40,10 +39,7 @@ char pushing_box_pos;
 
 char field[FIELD_SIZE];
 char field_original[FIELD_SIZE];
-char undo_buffer[UNDO_BUFFER_SIZE];
 char attempted_move_dir;
-short undo_moves_remaining;
-short current_undo_slot;
 char c,r,i;
 char* next_level;
 char* current_level;
@@ -140,9 +136,7 @@ void init_player() {
     pushing_box = 0;
     pulling_box = 0;
 
-    // Undo related
-    current_undo_slot = 0;
-    undo_moves_remaining = 0;
+    reset_undo();
 }
 
 void load_next_level() {
@@ -186,55 +180,6 @@ void move_barrel_off_of(char i) {
     } else {
         field[i] = field_original[i];
     }
-}
-
-void undo_buffer_push(char move) {
-    // The undo buffer is a bit of a ring buffer that can hold 256 entries
-    // We always increment the current undo slot (allowing it to wrap and
-    // overwrite old entries) but we cap the undo moves remaining at 255
-    // as to not allow undoing moves which have been overwritten
-    char index = (char)(current_undo_slot >> 1);
-
-    if (current_undo_slot & 1) {
-        char value = undo_buffer[index];
-        undo_buffer[index] = value | (move << 4);
-    } else {
-        undo_buffer[index] = move;
-    }
-
-    current_undo_slot++;
-    if (current_undo_slot >= MAX_UNDO_MOVES)
-        current_undo_slot -= MAX_UNDO_MOVES;
-
-    if (undo_moves_remaining < MAX_UNDO_MOVES)
-        undo_moves_remaining++;
-}
-
-// Returns 0xFF if the undo buffer is empty
-char undo_buffer_pop() {
-    char value;
-
-    if (!undo_moves_remaining) {
-        return 0xFF;
-    }
-
-    undo_moves_remaining--;
-
-    if (current_undo_slot == 0)
-        current_undo_slot = MAX_UNDO_MOVES - 1;
-    else
-        current_undo_slot--;
-
-    // The parity of current_undo_slot tells us which nibble to check
-    if (current_undo_slot & 1) {
-        value = undo_buffer[current_undo_slot >> 1] & 0xF0;
-        value >>= 4;
-    } else {
-        value = undo_buffer[current_undo_slot >> 1] & 0x0F;
-    }
-
-
-    return value;
 }
 
 void main_menu_loop() {
